Refactor step05_labeling: Consolidate implementation into single module
